mcp: carve the command registry out of facade.rs (pure move)
Review round 26's P1: facade.rs had grown past the ~3k non-test
budget. The declarative table and its vocabulary — RiskLane,
ValueKind, the spec structs, the positional/flag constructors, the
flag! macro, the COMMANDS const, and the two pure-table invariant
tests (lane/op consistency, path uniqueness) — move verbatim to
facade/registry.rs, re-exported from the old location so every
reference keeps compiling. The only non-move edits are the permitted
pub(crate) widenings on ValueKind, the spec structs, and their
fields. facade.rs keeps the planner (build_args + transforms),
dispatch substitution, resolution, help/docs rendering, and the
behavioral tests: 1,262 / 2,257 pre-test lines after the split.
